Sort shop list when open list
